python/bot/requirements.txt
discord.py
fastapi
uvicorn[standard]
pydantic
python-dotenv
APScheduler
aiohttp
audioop-lts
psycopg-binary
tortoise-orm[asyncpg]